- add WIP startup screen - off by enable flag for now - tweak default EnvironmentSettings values - hide 'Is Foreground scene' in editor - they all will be - terrain shader tweax

+++ Source/Engine/DotNET/Runtime/Editor/EditorActionStack.cs
@@ -5,7 +5,7 @@ namespace Hyperion
 {
     [ClassBinding(Name = "EditorActionStackState")]
     [Flags]
-    public enum EditorActionStackState : uint
+    public enum EditorActionStackState : byte
     {
         None = 0,
         CanUndo = 0x1,
